config: use atom defaults for missing local keys
When using a local config in projects, the linter assumes every key is
defined and has a value. Especially if "execPath" is not set in the
local config, linter-gcc trips and produces stack traces like the one
listed in #138. Fix things by loading the global Atom first and
replacing keys set via a local config afterwards.

Signed-off-by: Bert van Hall <bert.vanhall@gmx.de>